requirements_new.txt
fastapi
uvicorn
jinja2
sentence-transformers
faiss-cpu
sqlalchemy
aiosqlite
python-dotenv
pydantic
langchain
langchain-openai
langchain-community
langchain-core
transformers
python-multipart
apscheduler
pandas
datasets
huggingface-hub
